SetupGUIEditor_Commit
Followed labs exactly to set up GUI editor.  Yet to add to any controls
this way.

Not the autogenerated Editor source files have overwritten the manual
GUI code from previous commits.

Next stage is to finish the lab code (stereo panner) in this branch, and
experiment with the GUI editor.